+++ repolink/manual_deploy.go
@@ -1,4 +1,4 @@
-package repo_link
+package repolink
 
 import (
 	"gopkg.in/errgo.v1"
