+++ repolink/manual_review_app.go
@@ -1,4 +1,4 @@
-package repo_link
+package repolink
 
 import (
 	"gopkg.in/errgo.v1"
