+++ repolink/show.go
@@ -1,4 +1,4 @@
-package repo_link
+package repolink
 
 import (
 	"fmt"
